requirements.txt
streamlit
langchain_core
langchain
langchain-text-splitters
python-dotenv
ollama
langchain_community
openai
langserve
fastapi
uvicorn
sse_starlette
bs4
langchain_groq
groq
chromadb
requests
tiktoken
